Basic implementation of req resp logging
Logs info about any incoming requets and the headers of it's response.
