+++ upload-server/pkg/uid/uid.go
@@ -1,0 +1,18 @@
+package uid
+
+import (
+	"crypto/rand"
+	"encoding/hex"
+	"io"
+)
+
+func Uid() string {
+	id := make([]byte, 16)
+	_, err := io.ReadFull(rand.Reader, id)
+	if err != nil {
+		// This is probably an appropriate way to handle errors from our source
+		// for random bits.
+		panic(err)
+	}
+	return hex.EncodeToString(id)
+}
